// src/entities/Country/model/types/counrty.ts
export enum ECountry {
  RUSSIA = 'Россия',
  BELARUS = 'Беларусь',
  UKRAINE = 'Украина',
  KAZAKHSTAN = 'Казахстан',
  USA = 'США',
  GB = 'Великобритания',
  CHINA = 'Китай',
  CANADA = 'Канада',
  FRANCE = 'Франция',
  GERMANY = 'Германия',
  ITALY = 'Италия'
  // The rest of you, I'm sorry
}